Add PrivateRoute and UserPortfolioView components, secure UserPortfolioView
