+++ client/src/components/Navbar.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { Link } from "react-router-dom";
 import { useAuth0 } from "../authenticationWrapper.js";
 
 const NavBar = () => {
@@ -17,6 +18,13 @@ const NavBar = () => {
       )}
 
       {isAuthenticated && <button className="login-button" onClick={() => logout()}>Log Out</button>}
+
+      {isAuthenticated && (
+      <span>
+        <Link to="/">Home</Link>&nbsp;
+        <Link to="/portfolioView">User Portfolio</Link>
+      </span>
+    )}
     </div>
   );
 };
